+++ src/components/common/BaseInputField/index.ts
@@ -1,0 +1,1 @@
+export { default as BaseInputField } from './BaseInputField';
