+++ src/components/common/FooterButton/index.ts
@@ -1,0 +1,1 @@
+export { default as FooterButton } from './FooterButton';
