+++ src/components/common/MobileLayoutContainer/index.ts
@@ -1,0 +1,1 @@
+export { default as MobileLayoutContainer } from './MobileLayoutContainer';
