+++ src/components/features/payments/InputField/CardNumberInputField.tsx
@@ -1,6 +1,6 @@
 import { ChangeEvent, RefObject } from 'react';
-import BaseInputField from '../../../common/BaseInputField/BaseInputField';
-import Input from '../../../common/Input/Input';
+import { BaseInputField } from '../../../common/BaseInputField';
+import { Input } from '../../../common/Input';
 import { ERROR_TYPE_TO_MESSAGE, ErrorType } from '../config/error';
 import {
   CARD_NUMBER_INPUT_TYPE,
